Add shadcn
This took a bit of work. First, the setup is a bit different with
Tailwind v4. Second, we're setting it up in lib, which doesn't have
app oriented configs. Third, we're using workspaces (this didn't seem
to impact things as much).

To satisfy the init script, I had to add a dummy index.css, which is
a standin for Tailwind configuration. I also had to make sure that
all of the dependencies, including tailwindcss were present in
package.json.

The rest of the files were created by the init script (utils.ts, which
apparently needs to be in lib, components.json, and vite.config.ts).
I didn't update tsconfig.json, though the docs say to do so.